outsourced the LaunchRequestHandler
